+++ src/test/kotlin/omok/domain/judgement/ThreeJudgementTest.kt
@@ -1,6 +1,10 @@
-package omok.domain
+package omok.domain.judgement
 
-import omok.domain.judgement.ThreeJudgement
+import omok.domain.BlackStone
+import omok.domain.HorizontalAxis
+import omok.domain.Player
+import omok.domain.Position
+import omok.domain.WhiteStone
 import org.assertj.core.api.AssertionsForClassTypes.assertThat
 import org.junit.jupiter.api.Test
